+++ src/components/Container.ts
@@ -1,7 +1,0 @@
-import styled, { StyledComponentClass } from 'styled-components';
-
-const Container: StyledComponentClass<any, any> = styled.div`
-  background: white;
-`;
-
-export default Container;