+++ typings/react-motion.ts
@@ -1,0 +1,7 @@
+declare module 'react-motion' {
+  type Config = {
+    [key: string]: any;
+  };
+
+  export const spring: (num: number, config: Config) => void;
+}